executor/requirements.txt
ultralytics
opencv-python-headless
Pillow
numpy